+++ src/database/operations/files/loadFileIfCreatedBy.ts
@@ -1,0 +1,9 @@
+import { generateLoadItemOperation } from '../generators';
+import type { File, KeycloakId } from '../../../types';
+
+const loadFileIfCreatedBy = generateLoadItemOperation<
+	File,
+	[fileId: number, createdBy: KeycloakId]
+>('files.selectById', 'File', ['fileId', 'createdBy']);
+
+export { loadFileIfCreatedBy };